requirements.txt
python-dotenv~=1.0.1
jupyterlab
ipywidgets~=8.1.5
requests~=2.32.3
numpy~=1.26.4
pandas~=2.2.3
scipy
scikit-learn~=1.6.0
matplotlib~=3.10.0
gensim~=4.3.3
torch
transformers~=4.47.1
tqdm~=4.67.1
openai~=1.59.3
gradio~=5.9.1
langchain
tiktoken
faiss-cpu
langchain-openai
langchain_experimental
langchain_chroma
langchain[docarray]
datasets~=3.2.0
sentencepiece
matplotlib
google-generativeai
anthropic~=0.42.0
scikit-learn
unstructured
chromadb~=0.5.20
plotly~=5.24.1
jupyter-dash
beautifulsoup4~=4.12.3
pydub
modal~=0.71.0
ollama~=0.4.5
accelerate
sentencepiece
bitsandbytes
psutil
setuptools
speedtest-cli

ipython~=8.31.0
protobuf~=5.29.2
huggingface-hub~=0.27.0
pydantic~=2.9.2
joblib~=1.4.2